Initial commit
Added the build files for `Minecraft_FormUI_Reworked`.
A function named `CUSTOM_SHOW` is responsible to handle the forced UI to appear.
There are two new properties to each form type: `force` and `CLONE_SHOW`.
`CLONE_SHOW` is nothing more but a clone of the original function, while `force` defines if the UI should appear when UI is not busy anymore.